refactor(ui): split command renderer fixtures under the source length budget
renderer-fixtures.ts crossed 350 lines when the field-wiring fixture landed
on top of the drag-and-drop merge train. Move the command, history, and
shortcut fixtures into renderer-fixtures-commands.ts (mirroring the overlays
split) and teach the catalog renderer gate to read the new module.
