+++ npm/ui/core/scripts/renderer-fixtures.ts
@@ -1,6 +1,8 @@
+import { commandRendererFixtures } from "./renderer-fixtures-commands.ts";
 import { overlayRendererFixtures } from "./renderer-fixtures-overlays.ts";
 
 export const rendererFixtures = [
+  ...commandRendererFixtures,
   ...overlayRendererFixtures,
   {
     filename: "DragAndDropConsumer.vue",
@@ -67,87 +69,6 @@ const bravo = sortable.registerItem({ key: "bravo", element: second });
     <li ref="first" v-bind="alpha.itemProps" tabindex="0">{{ items[0] }}</li>
     <li ref="second" v-bind="bravo.itemProps" tabindex="0">{{ items[1] }}</li>
   </ul>
-</template>
-`,
-  },
-  {
-    filename: "CommandConsumer.vue",
-    source: String.raw`<script setup lang="ts">
-import { ref } from "vue";
-import { useCommandRouter } from "./command.ts";
-
-const output = ref("");
-const router = useCommandRouter<"save">();
-router.register({
-  id: "save",
-  title: "Save Document",
-  run: () => {
-    output.value = "saved";
-  },
-});
-</script>
-
-<template>
-  <button type="button" :disabled="!router.isEnabled('save')" @click="router.execute('save')">
-    Save
-  </button>
-  <output>{{ output }}</output>
-</template>
-`,
-  },
-  {
-    filename: "HistoryConsumer.vue",
-    source: String.raw`<script setup lang="ts">
-import { ref } from "vue";
-import { useHistory } from "./history.ts";
-
-const value = ref(0);
-const history = useHistory();
-function increment() {
-  const before = value.value;
-  value.value += 1;
-  history.pushSnapshot({
-    before,
-    after: value.value,
-    apply: (next) => {
-      value.value = next;
-    },
-    label: "Increment",
-  });
-}
-</script>
-
-<template>
-  <button type="button" @click="increment">Increment</button>
-  <button type="button" :disabled="!history.canUndo.value" @click="history.undo()">Undo</button>
-  <button type="button" :disabled="!history.canRedo.value" @click="history.redo()">Redo</button>
-  <output>{{ value }}</output>
-</template>
-`,
-  },
-  {
-    filename: "ShortcutConsumer.vue",
-    source: String.raw`<script setup lang="ts">
-import { ref } from "vue";
-import { formatShortcut, useShortcutRegistry } from "./shortcut.ts";
-
-const host = ref<HTMLElement | null>(null);
-const count = ref(0);
-const registry = useShortcutRegistry({ target: host, platform: "standard" });
-registry.register({
-  shortcut: "Mod+K",
-  description: "Open the palette",
-  handler: () => {
-    count.value += 1;
-  },
-});
-</script>
-
-<template>
-  <div ref="host" tabindex="0" :data-pending="registry.pendingSequence.value.length || undefined">
-    <kbd>{{ formatShortcut("Mod+K", { platform: "standard" }) }}</kbd>
-    <output>{{ count }}</output>
-  </div>
 </template>
 `,
   },
